requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
confluent-kafka==2.10.1
curl_cffi==0.11.3
Deprecated==1.2.18
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cache2==0.2.2
flake8==7.2.0
frozendict==2.4.6
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
limits==4.2
mccabe==0.7.0
multitasking==0.0.11
numpy==2.0.2
packaging==24.2
pandas==2.3.0
peewee==3.18.1
pendulum==3.1.0
platformdirs==4.3.8
pluggy==1.6.0
protobuf==6.31.1
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
pytest==8.4.0
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==4.6.0
requests==2.32.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
structlog==25.4.0
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2
yfinance==0.2.63
